Migrate relay pagination tests from mirage to relay wrapper
